@@ -16,8 +16,8 @@
     "extract-text-webpack-plugin": "^3.0.2",
     "history": "^4.9.0",
     "mdbreact": "^4.17.0",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-images": "^0.5.19",
     "react-loader-spinner": "^2.3.0",
     "react-redux": "^7.0.3",
